backend/requirements.txt
fastapi==0.74.1
uvicorn==0.17.5
python-cas==1.5.0
itsdangerous==2.1.2

# dev
pytest-cov==3.0.0
pytest==7.0.1
pylint==2.12.2
requests==2.26.0
